Update patch set 1
Patch Set 1:

After applying this patch, nothing changed for me. Both themes still give the errors and the errors are still reversed from how they were before: https://www.dropbox.com/s/kowy4xm0szsmx0r/themes.jpg

It's not that I mind the errors so much now that you can use the Apply Anyway option, but it's still a mystery as to why the errors are reversed. It changed between 11/23 and 11/25.

Patch-set: 1